Allow editing of opening date for categories
Up to the recent change in the transaction editor to verify that all
opening dates of all referenced accounts/categories are prior to the
transaction date, the opening date of a category did not play a role
and was not editable. It was set to the date the category was created.

This creates a problem, in case I want to enter a transaction with
post date yesterday using a new category I created today.

Therefore, we simply make the field editable. Also, the standard
opening date for categories is now the first day of the current
fiscal year instead of the current date.