Don't modify transaction date automatically
The recently introduced consistency check feature to modify the
transaction date to always make sure to be post the opening date of any
referenced account may cause some serious trouble when operating on
historical data entered using previous versions of KMyMoney. This could
have the effect that transactions get a post date even in different
years. This certainly causes severe problems with the financial figures.

This change prohibits the automatic change of a transactions post date.
In case the date problem is caused by a category or stock account, the
opening date of the account is modified to the post date of the
transaction. In case an account is causing the problem, the user will
receive information about the name of the account and the transaction in
question. It is left to the user to fix the problem manually.

In rare cases, the problem cannot be fixed since by modifying the
opening date of the account, as the current UI does not support to
change it. This is true for e.g. loan accounts.

This implementation operates a lot safer on user data than the original
one. I had reviewed the code and found it to be OK, but did not see the
side effects that I found while running the consistency check on my
production data.

(cherry picked from commit af7939608b3aa305a86b0ad6196beacb52c63108)