Don't change the original start date of a schedule
Entering schedules updates the next due date of the schedule. Before
this change, this also used to change the start date member variable
which is the reason that the day of the first execution of the schedule
is lost over time when it is modified because it is in the range of
29-31 and the month for the transaction does not have that many days. In
this case, the day was adjusted and caused future transaction not to
happen on this day anymore.

I have updated testcases and added a new one but still we need more
testing before we cherry pick this change over to the 4.7 branch.

BUG: 345061