Fix transaction purpose download for SEPA tags
Since a few institutions have their very special interpretation of
specification, the transaction importer has been enhanced to deal with
this crap. On top, this requires at least an AqBanking version as of
today. Thanks to Martin Preuss for providing the necessary changes in
AqBanking.

If one uses this code to retrieve transactions that already have been
downloaded into KMyMoney, duplicates will be generated because the
contents of the memo field changes with this code.