Combined the questions if a transaction should be matched
Moved the question from TransactionMatcher::match to
MyMoneyStatementReader::askUserToEnterScheduleForMatching. Added
MyMoneyTransaction & importedTransaction parmeter to
MyMoneyStatementReader::askUserToEnterScheduleForMatching, as
importedTransation is required to calculate if the question from
TransactionMatcher::match should be shown.

This deactivates the question on manual merge which sould get fixed
later.

The original review request was for KMyMoney's Qt4 version.
Christian David converted it to a Qt5/KF5 and did some minor
changes.

REVIEW: 126875