Changed way of saving files which should fix some bugs
The save operation seems to fail every time - it never changed the
orginal file and never reported any issues. I did not find the exact
reason for this bug but I am quite sure it was caused by an incorret
usage of QSaveFile (under some circumstances close() instead of commit()
was called).

Now KMyMoney creates its own temporary file to write to (if needed).
This also works using KGpgFile, which should fix Bug 356399.

The remove() and rename() operations are not atomic which is not so
good as this could result in dataloss if the first operation fails.
However, this is the best OS independet process I could find.

Errors during writing of compressed files may not be detected. I think
this issue should be fixed upstream.

BUG: 356399
FIXED-IN: 5.0
REVIEW: 127108