Don't insert linefeeds between chunks of purpose field
The purpose field is transferred in multiple chunks between AqBanking
and the application. This might happen in the middle of a word. This has
been changed to keep the purpose as is and insert linefeeds if certain
fields are detected (e.g. EREF:, CREF:, MREF: etc.) The algorithm to
construct the transaction hash to detect duplicates has not been
changed to remain backward compatability.