Fixed new file dialog
The check for the filename did not work correctly. Also the conversion
QUrl -> QString -> QUrl after the dialog is finished always failed. The
solution is to not convert the QUrl to a string.

Now deprecated functions are (still) used and the user can select an
existing file in the dialog even if the widget forbids it. Both issues
should be addressed later.