Change KDialog to QDialog in any .ui files that used KDialog.
KDialog does not work well with Qt Designer.

In most of the dialogs the only change required was to replace
KDialog with QDialog in the .ui file.  The find-transaction
dialog also required removing the KDialog button declarations
from the .ui file and adding a call to setButtons in the
constructor.

REVIEW: 127680

Signed-off-by: Thomas Baumgart <thb@net-bembel.de>