Reduce the precision of shares and values
The problem described in bug 345655 is caused by a fractional part of the
splits value that is larger than the fractional part of the accounts currency.
This leads to the effect, that amounts are displayed with correctly
but the underlying amount used for calculations is not equal to the
amount displayed. This will end up in KMyMoney presenting the user
strange messages like 'Unassigned value of 0.00' which do not make
sense. The difference is in reality between 0 and 0.01, e.g. 0.008.

BUG: 345655
FIXED-IN: 4.8.0
REVIEW: 128061