Renamed SQL column "order" to "userOrder" in some tables
The name "order" is a reserved keyword. It can be used as identifier but
only if it is escaped. Unfortunately the escape sequence is database
dependent.

This patch is untested as another bug prevents me from opening a
database. Scince this fix is kind of urgent, I publish it anyway. The
port to 4.8 should be straight forward.

BUG: 365615
FIXED-IN: 5.0