Detect if libalkimia is too new
KMyMoney fails to build with libalkimia version 6.0.0 and larger. This
ended up in using the wrong version of moc if Qt4 and Qt5 development
tools were installed on a system or a failure that Qt5Core was missing
if Qt5 was not installed.

We now detect the wrong version and give a clear error message during
cmake run.