Added client uid support for OFX
Added a slightly modified and improved version of the patch which
is attached to bug #366326 and adds support for UI elements to enter a
clientuid and send it to LibOFX if it is supported there. In case the
LibOFX version available does not support clientuid the widget will be
shown disabled.

CMake detects the presence by checking the LibOFX header file to find
the clientuid field in the respective structure.

BUG: 366326