Renamed SQL column "order" to "userOrder" in some tables
The name "order" is a reserved keyword. It can be used as identifier
but only if it is escaped. Unfortunately the escape sequence is database
dependent. Added upgrade feature to convert database on open.

Accomplished compatability with DB layout version 10 as used in master.

This has yet been tested against SQLite but not MySQL.
BUG: 365615
FIXED-IN: 4.8.1