Moved plugin loader out of kmm_plugin and put it into the executable
This patch ported the plugin system to KF5.

The "plugin" shared library is for plugins now and does not include the
loading of plugins (which was kind of wired). All loading operations of
standard plugins is done in KMyMoneyApp now.

The interface system for plugins needed do be reworked to
great degree. Some overly complex operations were replaced by simpler
ones. All plugins were adopted to that. Also it is more type safe now.

Now the user configuration which plugins should be loaded or not is
respected (standard plugins only).

An old bug was fixed: If the configuration in the plugin settings
changed the plugins are loaded/unloaded after the user accepts the
dialog (perviously this was done when the checkbox was clicked).

The new system is still not finished. However, it is a great step in
the right direction. An uncomplete list of FIXMEs (or bugs?):

1) pluginloader.{h,cpp} are gui only now and should be merged into the
settings dialog.
2) The UI for plugins still has some issues, e.g. the online banking
plugins which are loaded on demand (and are not deactivatable) have a
checkbox for loading and unloading which has no function.
3) Unloading plugins may not work correctly (I think it never did).
4) The payeeIdentifiers are not listed in the plugin settings page.

REVIEW: 129393