Fix the report stylesheet and export.
While migrating from KDE4 it is possible to end up with a css file
setting which does not exist anymore so check the file's existence
before returning it as valid.

The export functionality was using KFileDialog to add a checkbox that
would control css inclusion. Now QFileDialog is used and the css is
always included. The related dead code was removed. An alternative
would be to use KFileWidget inside a QDialog but that code would be
too verbose and the value gained would not worth it (I'm not even
sure what can be achieved by not including the CSS but it can be moved
as a global option if it is necessary).