Add exact payee name match option
The current 'Match on payees name' option works on a partial match
of the name. In some scenarios, an exact match is expected. The
new option allows to select that on a per payee basis.

The default for new payees entered through the GUI is not changed.
It remains 'No matching'. The default for new payees created during
statement import is the new 'Match exact name' option.

Also, the radio button based selection has been changed to a combobox.

GUI:
CCBUG: 374352