More work on the new ledger code
This now adds a first working version of the new ledger code.
It's not feature complete compared to the old code, but one could
open multiple accounts in parallel, browse transactions in them,
edit regular transactions and also split transactions.

The amount entry has been changed to always have two widgets. This
reduced the complexity a lot, esp. in the split editor. I am not
sure if I got all the credit and debit right, so please take a look.

Multi currency support is not yet added as well as support for
investment transactions. Also missing are the fancy delimiters
in the ledger.

Please don't use it on your production data as it might still have
some areas that are untested.