Add support to use a dedicated opening balance account.
With recent implementation kmymoney is not able to use a
dedicated account for opening balances. Instead it (creates)
and uses an opening balance account with the translated name
of "Opening Balances" which may be language specific.

To specify an opening balance account to be used on adding new
accounts one way of operation is to set the "opening balance
account" flag in the account editor for equity accounts.

This is enabled if no other account contains this flag. Also
changing the state of the flag is only possible if no transactions
are assigned to the account having this flag.

Another way to use this support is to flag an opening balance
account in a kmymoney account template file.

On creating a new kmymoney file the "opening balance account"
flag is imported from a template account flag if specified
in the following form:

<account type="16" name="9000 Saldovortragskonten">
   <flag name="OpeningBalanceAccount"/>
</account>

The template admin needs to make sure that only one template account
has this flag set.

Exporting the current kmymoney file to an account template exports
this flag too.

If there is at least one dedicate opening balance account in a kmymoney
file the name will be used as template for adding additional currency
related opening balance accounts in the same hierachy level as the
referenced opening balance account. On the search all accounts having
a currency postfix '(XXX)' are ignored.

BUG:370290
REVIEW:130143