Fixed a bunch of warnings
Most of them were just marking fall throughs in switch/case statements
so that the compiler does not spit out a warning where we wanted them.
One break statement was missing. Also, some date adjustments in reports
did not work as expected since the return value was not assigned back to
the variable.