Defer loading of views
Purpose of this patch is to speed up KMyMoney startup. Payees view and
Budgets view have QSpliter which is relatively costly to initialize.
Should user use this pages, they will be loaded dynamically.

Changes:
1) View frames have equal type, so it's shorter to have hash table than
separate pointers and viewID enum can be reused as well.
2) UIs are inherited as private and not public for correctness.
3) Constructors has been moved to init() methods which will be executed
when page is about to be shown.

Differential Revision: https://phabricator.kde.org/D6465

Signed-off-by: Łukasz Wojniłowicz <lukasz.wojnilowicz@gmail.com>