Return saves updated price information
When the focus was in the value edit field, the Cancel button for some
reason was the default button. So pressing Return caused the cancelation
of the entry and the values were not updated. The original code therefor
had Ctrl-Return working. The change fixes the bahavior so that Return
alone saves the data and behaves as clicking OK with the mouse.