Fix crash accessing wrong data array on investment price graph
The crash happened because position.column is not used to check
the requested data array instead it checks always the first data
array which fails later in case position.row exceeds array
size.

This commit needs to be posted to the KDChart bug tracker too to be
included in further KDChart releases.

FIXED-IN:4.8.1
BUG:383160