Fix 'Broken net worth graph in case data contains only the same value'
If chart model contains only data with the same value KDChart 2.6
automatically adjust the vertical range to a non visible area
and displays x axis labels at the top of the diagram.

The fix is to disable "auto adjust vertical range to data". The used
value has been empirically determined.

FIXED-IN:4.8.1
BUG:383161