Fix 'Crash on investment charts with multiple investments'
The root cause of this crash is an implementation bug in
CartesianDiagramDataCompressor::slotColumnsAboutToBeInserted().

In case m_model is set, adding columns to m_data requires to add a
QVector<DataPoint> instance sized to the model row count. Instead the
size is limited to m_xResolution which depends on the diagram zoom factor
and other values. Having a too small sized column in m_data let
slotRowsAboutToBeInserted() crash.

This patch needs to be pushed upstream to have it in further kdiagram
releases.

Fixup of commit 7516308b.

BUG:384021
FIXED-IN:4.8.1