Fix KDE CI Windows builds
Anything hosted on git.kde.org goes into a non-standard prefix (ie.
not /usr or /usr/local). It's made available to CMake through
the usual variables though.

The issue is that $prefix/include probably isn't added by the KF5
CMake files, but $prefix/include/KF5 will be added.

On most systems, which use a standard prefix like /usr this works
fine as other packages ensure /usr/include/ is in the compiler's
search path. It fails on the CI though as it uses a different prefix
(which is a perfectly supported scenario with Frameworks)

CCMAIL:bcooksley@kde.org