Port WebConnect feature to KF5
This implements the WebConnect feature to import a file directly from
another application (e.g. web browser) simply by starting KMyMoney and
providing the downloaded data in a file via command line argument.

In case KMyMoney knows the format of the file and another instance of
KMyMoney is already running it is asked to import the data into the
opened file.

In case no KMyMoney instance is running, the last used file is opened to
import the data.

Currently. the OFX plugin supports the WebConnect feature.