Revert/improve last commit
The cause to the problem why KMyMoney did not pick up the CSS file on my
system was that QStandardPaths::standardLocations() returned at least
three directories. The last one in the list does not contain the files
the logic requires and thus the files in the application config
directory never got created.

The new logic scans all source directories and stops on the first that
contains all CSS files we try to theme here.