Fix 'Wrong Currency conversion rate in Reports'
The root cause was that the conversion of reports value to the base
currency is done on grid entry date and not on transaction date.
The price on grid entry date may differ from the price on the
transaction date and results in unexpected sums.

BUG:385857
FIXED-IN:4.8.2

Test Plan: - generated with appended test case

Reviewers: #kmymoney, wojnilowicz

Reviewed By: wojnilowicz

Subscribers: wojnilowicz

Differential Revision: https://phabricator.kde.org/D8678