Free KBudgetView from KMyMoneyApp
KBudgetView implementation is fragmented. Part of it is in KMyMoneyApp
where only KBudgetView made calls to it. That cuases KBudgetView to
include big KMyMoneyApp totally unnecessary. This patch fixes it.
In case KMyMoneyApp would like to call KBudgetView methods in future,
then it should do it through signal/slots in KMyMoneyView, which will
preserve encapsulation of KBudgetView.

Differential Revision: https://phabricator.kde.org/D9278