Fix SEPA plugin installation
Summary:
add_library() may add a "lib" prefix to the plugin's filename, which causes it not to be found.
Using kde4_add_plugin() instead again fixes that.

This change to add_library() was (accidentally?) backported from the KF5 version in commit c78183f18201 .

BUG: 386547
FIXED-IN: 4.8.2

Intended for the 4.8 branch only, of course.

Test Plan:
The plugin is again installed as konlinetasks_sepa.so here (like in 4.8.0, and also the KF5 port).
The error message that the plugin could not be found does not occur anymore, kmymoney prints

> onlineTask available "org.kmymoney.creditTransfer.sepa"

 to stdout now.

Haven't tested the actual functionality though.

Reviewers: #kmymoney, habacker

Reviewed By: #kmymoney, habacker

Subscribers: tbaumgart

Tags: #kmymoney

Differential Revision: https://phabricator.kde.org/D9545