Specify the config file before construction time
It seems that KMyMoneySettings aren't as shared as I thought. It seems
that call to KMyMoneySettings in every translational unit uses different
instance.
Remaining at old code would require to create instance of config file in
every translational unit in which KMyMoneySettings is used and not only
in main function. That misses the point, so config filename is hardcoded
as before in kcfg file.

BUG:388640
FIXED-IN:5.0