Optimize MyMoneyTransactionFilter for speed
Opening home page on startup with big file (mine = 6 MB) takes lots of
computational resources and with it time.
This patch reduces need for that resources by 70 % in my test scenario
in debug mode.
This was possible after removing bottleneck that has been identified to
be in MyMoneyTransactionFilter and the main cause was copying of
MyMoneySplit objects.
Instead of filling matchingSplits list in MyMoneyTransactionFilter right
from the start and then removing unneeded ones, no split is added or it
is added if it qualifies.

Changes:
1) "container.isEmpty()" instead of "container.count > 0", as it's
faster,
2) QVector instead QList for MyMoneySplit, as it's faster,
3) constify containers before for-range-loop, as it avoids costly
detaches,
4) reduce unneeded ifs-trees,
5) replace QMutableListIterator with faster for-range-loop,
6) use QStringLiteral instead of const char* for split actions, to avoid
costly memory allocations,
7) use QHash::contains instead of QHash::find for readability,
8) remove m_matchingSplits variable in favor of local ones, as it was
short lived anyway,
9) new hackish m_matchOnly variable, to not fill matchingSplits list if
we don't need it.

Differential Revision: https://phabricator.kde.org/D9856