Fix the Austrian Schilling symbol problem
The symbol for the Austrian Schilling was not correctly encoded. Since
it is loaded into the data file once and not touched anymore it is hard
to spot. This patch fixes the problem on the fly for existing files if
you open the Tools/Currencies dialog and also for new files from the
beginning.