Fix column visibility when turned on for single view
In case multiple account/category views are loaded, and the user adds a
column to the view he is asked if the column should be added to all open
views. In case he answered no, the column was not visible. This change
fixes that so that the column is visible in this case.