Silence CMake warning about policy CMP0071
Summary: This policy was introduced in CMake 3.10 and will annoy users until we set that CMake version as minimum. Probably we do not require this feature, yet. It is only set to old, to prevent accidential use of the new behavour. If the new behaviour becomes important, cmake_minimum_required has to be set to "3.10".

Test Plan: Run `cmake .`

Reviewers: #kmymoney, wojnilowicz, tbaumgart

Reviewed By: #kmymoney, wojnilowicz, tbaumgart

Subscribers: tbaumgart

Differential Revision: https://phabricator.kde.org/D10162