Fix issues with multiple instances of KMyMoneySettings
Calling a static method from class KMyMoneyGlobalSettings inside
kmm_widgets library surprisingly creates an additional KMymoneySettings
instance, which does not reflect the recent state from the setting dialog.
The reason for this behavior is caused by duplicated implementations of
class KMymoneySettings (kmm_config, settings) and embedding the static
library kmm_config into shared libraries (kmm_mymoney, kmm_widgets and
kmm_ofximport). Each embedded copy has it's own instance of class
KMymoneySettings.

The solution for this issue, which is done with this patch, is to remove
the duplicated implementation from the settings library and to make
kmm_config a shared library.

This patch also renames kmm_config to kmm_settings and place it into it
own subdirectory to keep in sync with 5.0 branch.

CCBUG:374123
CCBUG:389631
FIXED-IN:4.8.2

Test Plan: compiled on linux

Reviewers: #kmymoney, tbaumgart

Reviewed By: tbaumgart

Subscribers: tbaumgart

Differential Revision: https://phabricator.kde.org/D10322