Remove online job view if not needed
Summary:
In case no online job plugin is available, the online job view is not
needed. This change removes the view from the application in case no
online plugin is active at startup of the application.

BUG: 377760
FIXED-IN: 5.0.1

Test Plan: Disable KBanking and OFX plugin and check that page is not shown. Enable at least one of them to have the page back after the next application start.

Reviewers: wojnilowicz

Reviewed By: wojnilowicz

Differential Revision: https://phabricator.kde.org/D10457