Respect currency symbol position of locale
QLocale does not provide information about the position of the currency
symbol with respect to the amount. This used to be the case in KDE4
where KLocale provided it. As part of porting, this was set to a
constant which might not be correct in certain locales.

This commit provides logic to pull out the information from the locale
on plattforms that support it. For others, it is still set to a fixed
value.