Move opening files to KMyMoneyApp
In my opinion opening file has little to do with views, so I moved it to KMyMoneyApp. KMyMoneyApp in many places already called KMyMoneyView in this regard, so the move around seems like a good choice.
It also serves as a one step forward in making KMyMoneyView independent of KMyMoneyApp.
The next step would be to reduce a rather long code to open a simple file.

Differential Revision: https://phabricator.kde.org/D10790