Set AutoFillTransaction to correct value
The AutoFillTransaction setting never received a value other than zero.
This was caused by a non standard layout in the group box of the setting
which apparently caused the kcfg_ magic to fail. This caused the dialog
with previous transactions not to be shown.

Using only QRadioButtons inside the QGroupBox fixes this problem.

BUG: 391048
FIXED-IN: 5.0.1