Introduce KTransactionFilter
Purpose of this patch is to make KReportConfigurationFilterDlg independent from KFindTransactionDlg. It's done by introducing KTransactionFilter which doesn't depend on KMyMoneyRegister and is a kind of frontend for setting MyMoneyTransactionFilter.
This approach will allow us to make plugin out of KReportsView and it makes interface tidier as presented on the image below.

Differential Revision: https://phabricator.kde.org/D11014