Convert enumeration MyMoneyReport::EColumnType into a class to be more type safe
Type safety is increased by moving the enum into a separate name space (here
a class is used because namespaces are not allowed inside classes).

Q_ENUM(s) or QFlags could also not be used because the related enums need
to be defined somewhere outside, which does not solve the issue having all
constants in MyMoneyReport namespace.

CCBUG:391091

Test Plan: compiled on linux

Reviewers: #kmymoney

Reviewed By: #kmymoney, tbaumgart

Differential Revision: https://phabricator.kde.org/D11043

Signed-off-by: Ralf Habacker <ralf.habacker@freenet.de>