Refer to views by KMyMoneyViewBase
I want to refer to the views in a generic way in order to be able to manage them in a simple way i.e. through viewFrames and viewBases.
It's not possible fully right now, so many static_cast appears, but they will disappear with an advent of reworked signals framework.

Differential Revision: https://phabricator.kde.org/D11464