Extract IBAN, BIC and currency from MyMoneyAccount object
Extracting IBAN and BIC from the MyMoneyAccount object was turned off
since it causes trouble on MSVC but it works OK on Linux. So I keep it
off for Q_OS_WIN only.

Also, the currency can be extracted from the MyMoneyAccount object in
case it is known. Otherwise, the base currency will be used.