Pluginize KOnlineJobOutbox
This patch makes online job outbox view optional instead of mandatory.
User can switch on and off this view and have more space on views pane.
Developers will benefit by saving on build time of this plugin.

It compiles KMyMoneyAccountCombo and links to static models library just
like QIF exporter does. Something has to be done with this, but for now
switchable outbox is enough feature to be committed.

Differential Revision: https://phabricator.kde.org/D11841