Pluginize XML Storage
The goal of this patch is to open supported files in alike ways i.e.
through KMyMoneyPlugin::StoragePlugin. This is the first step to make
Qt5::Xml dependency optional by encapsulating saving as XML file in a
plugin.
The patch is transient as "opening", "saving" and "saving as" could be
made consistent in kmymoney.cpp. GNC reader could also be made
consistent with XML and SQL storage. KMyMoneyPlugin::StoragePlugin could
also be extended by identification of its own file formats.
With this patch I want to know if all works well.

Differential Revision: https://phabricator.kde.org/D13279