Enable foreign keys support in SQLite
Deleting from kmmOnlineJobs leaves kmmSepaOrders in tact because there
is no:
1) ON DELETE relation between kmmSepaOrders and kmmOnlineJobs,
2) foreign_keys is disabled on default on SQLite, which makes ON DELETE
possible.

Leaving kmmSepaOrders in tact prevents saving new records to it, because
there aren't overridden automatically and thus regarded as non unique,
which fails the requirement of INSERT.
We can easily recreate kmmSepaOrders (with ON DELETE constraint) without
taking its content into account, as it would be deleted anyways.

Differential Revision: https://phabricator.kde.org/D13295