Move XML read methods to storage plugin
This is a step in XML encapsulation direction. Because overall it's a big undertaking, only read methods from mymoneystoragexml has been moved. Save methods will follow later. Because it's an error prone undertaking, old read methods have been left untouched for a quick and transparent testing. MyMoneyObjects are loaded in mymoneystoragexml as follows:
*1) using old way (constructor accepting QDomElement),
2) using new way (static read methods in mymoneystoragexml),
*3) returned object from old and new way are compared and inequality is signaled in terminal,
*4) returned object from old way is discarded and object from new way is going further,

*only for testing, will be removed with underlying code in the accepted patch

Differential Revision: https://phabricator.kde.org/D13831