Prepare for XML storage separation
Goal of this patch is to facilitate separation of XML saving and reading methods to XML storage plugin by:
1) adding new constructor for MyMoneyObjects

The new constructor accepts id, so an MyMoney object will be created in XML storage plugin at the spot and not by a copy from another object. That's one constructor more for every MyMoneyObject, but there will be one constructor (the one with QDomElement) less after complete encapsulation of XML support in KMyMoney

2) moving all remaining enums to mymoneyenums

Some enums were changed to enum class, so type safety is enhanced a little bit.

Differential Revision: https://phabricator.kde.org/D13581