Prevent crash on destruction of KTagContainer
I could happen, that the connected KMyMoneyMVCCombo widget looses the
focus and as part of it emits lostFocus() before the connection was
removed as part of the destruction in the Qt main loop but the
KTagContainer was already deleted. This causes a crash of KMyMoney as
the KTagContainer object does not exist anymore.

This change removes the connection as part of the KTagContainer
destruction so that the above mentioned scenario cannot happen anymore.