Provide timezone offset for OFX import
Summary:
In certain constellations the post date could be off by one day. This
can happen, when your institution provides timestamps that differ from
the one of your local time. This change allows to adjust the timestamps
during import so that the date will be shown correct.

The new setting can be used to adjust the timestamps of the transactions
during import to reflect your local time. It does not affect previously
imported transactions. The offset is formatted as HH:MM.

If you are located west of your bank's server than enter a negative
time, in case you are located east of it enter a positve time.

If you are not sure, leave this setting at 0:00.

See also: https://forum.kde.org/viewtopic.php?f=69&t=152300

{F5887276}
This shows the new setting with an offset of -3 hours.

Test Plan:
Requirement is to have an OFX server in different timezone.
Make sure to import a transaction that shows the problem of a post date
off by one day.
Delete the transaction in question.
Setup the correction value in the new account online setting and import.
Repeat the import and verify that the post date is now correct.

Reviewers: ostroffjh

Differential Revision: https://phabricator.kde.org/D13293