Move reports drawing to reports view plugin
This patch completes encapsulation of reporting features in KMyMoney. If a report is needed then it should be returned as:
1) QWidget
It will then be possible to show report outside of KReportsView and handle it completely inside KReportsView. This approach will allow us embedding any report anywhere with configuration controls built in a report.

2) QString
Some reports aren't graphical charts, so they need textual representation. Disadvantage of this approach is that chart cannot be handled inside KReportsView.

Differential Revision: https://phabricator.kde.org/D13687