Add field for institution's website URL
Summary:
This change adds an edit field for the URL of a bank's website to the
institution dialog. For security reasons, allow the https protocol is
used. The protocol does not need to be entered as part of the URL and in
case it is, it will be removed from the URL part. A button right to the
edit field allows to open the URL entered in the browser.

Whenever the URL text changes, a check will be started to verify that
the URL entered accesses a web server. In case it does, the button is
enabled otherwise disabled. In case a favicon is found at the URL it is
placed on the button otherwise, the standard KMyMoney icon for
institutions is used. In case of an invalid URL the icon will be removed
from the button.

No check is performed that it is really your bank's website that opens
when pressing the button.

{F5982397}
Button with a favicon drawn from the website.

{F5982404}
Same area with an account that is not managed by an institution.

{F5982419}
And an institution that does not provide a favicon, so we use our own.

Test Plan:
Added false URLs, URLs that do not provide a favicon and ones that
provide a favicon. Saved data to file and checked that the KVP
entries are filled.

Reviewers: #kmymoney, ostroffjh, wojnilowicz

Reviewed By: #kmymoney, wojnilowicz

Differential Revision: https://phabricator.kde.org/D13707