Only show a single summary of statement import
Summary:
Depending on the number of statements imported via online connects to
banks, the sheer number of dialogs popping up is annoying and time
consuming. This change reduces this to a single dialog which is also
shown if no transaction has been imported at all to notify the
termination of the process to the user.

BUG: 396886

Test Plan: I ran updates for single accounts as well as multiple accounts

Reviewers: wojnilowicz

Reviewed By: wojnilowicz

Differential Revision: https://phabricator.kde.org/D14439