Use standard colors with better readability
Summary:
The background colors used for imported and matched transactions is not
readable. This change sets the colors to be more readable. To further
differentiate between Imported and Matched transactions, the matched
transactions are now using bold text.

BUG: 397023

The first row shows a matched transaction, the second row an imported one

prior to change it looks like this:
{F6176182}

after the change it looks like this
{F6176190}

Reviewers: antoine

Reviewed By: antoine

Subscribers: antoine

Differential Revision: https://phabricator.kde.org/D14623